moved all utils in main file
